feat: add family

+++ backend/auth/views.py
@@ -6,7 +6,7 @@ from rest_framework.authtoken.models import Token
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.permissions import IsAuthenticated
 from .models import User
-from .serializers import UserSerializer
+from .serializers import UserSerializer, FamilySerializer
 
 """ 
 [signup api]
@@ -17,10 +17,18 @@ from .serializers import UserSerializer
     last_name: char
     phone_number: char
     birth: YY-MM-DD
-    is_senior: 0(unknown)/1(male)/2(female)
+    is_senior: true/false
 [login api]
     username: char
     password: char
+[create-family api]
+    family_name: char
+    senior_gender: 0(unknown)/1(male)/2(female)
+    senior_diseases: char
+    senior_interests: char
+[join-family]
+    family_id: uuid(6)
+    role: char
 """
 
 @api_view(['POST'])
@@ -35,30 +43,61 @@ def signup(request):
             "token": token.key
         }
         return Response(data, status=status.HTTP_201_CREATED)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    data = {"error": serializer.errors}
+    return Response(data, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['POST'])
 def login(request):
     authenticate_user = authenticate(username=request.data['username'], password=request.data['password'])
-
     if authenticate_user is not None:
         user = User.objects.get(username=request.data['username'])
         serializer = UserSerializer(user)
-        response_data = {
-            'user': serializer.data,
-        }
+        data = {'user': serializer.data,}
         token, created_token = Token.objects.get_or_create(user=user)
         if token:
-            response_data['token'] = token.key
+            data['token'] = token.key
         elif created_token:
-            response_data['token'] = created_token.key
-
-        return Response(response_data, status=status.HTTP_202_ACCEPTED)
-    return Response({"error": "user not found"}, status=status.HTTP_400_BAD_REQUEST)
+            data['token'] = created_token.key
+        return Response(data, status=status.HTTP_202_ACCEPTED)
+    data = {"error": "user not found"}
+    return Response(data, status=status.HTTP_404_NOT_FOUND)
 
 @api_view(["GET"])
 @authentication_classes([TokenAuthentication])
 @permission_classes([IsAuthenticated])
 def logout(request):
-    request.user.auth_token.delete()
-    return Response({"message": "logout was successful"})
+    if request.user:
+        request.user.auth_token.delete()
+        data = {"message": "logout was successful"}
+        return Response(data, status=status.HTTP_200_OK)
+    data = {"error": "user not login"}
+    return Response(data, status=status.HTTP_400_BAD_REQUEST)
+
+@api_view(['POST'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
+def create_family(request):
+    serializer = FamilySerializer(data=request.data)
+    if serializer.is_valid():
+        user = request.user
+        serializer.create(user=user)
+        data = {
+            "family": serializer.data,
+            "role": "senior"
+        }
+        return Response(data, status=status.HTTP_201_CREATED)
+    data = {"error": serializer.errors}
+    return Response(data, status=status.HTTP_400_BAD_REQUEST)
+
+@api_view(['PUT'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
+def join_family(request):
+    try:
+        serializer = UserSerializer()
+        serializer.join(request.user, request.data)
+        data = {"message": "join was successful"}
+        return Response(data, status=status.HTTP_200_OK)
+    except Exception as e:
+        data = {"error": e}
+        return Response(data, status=status.HTTP_404_NOT_FOUND)
